opik/streamlit/call-summarizer/pyproject.toml
[tool.poetry]
name = "call-summarizer"
version = "0.1.0"
description = "A Streamlit app for summarizing and categorizing call transcripts using LangGraph and Opik"
authors = ["Francisco <22344801+fschlz@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "call_summarizer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.32.0"
opik = "^1.7.26"
langchain = "^0.3.25"
langgraph = "^0.4.5"
langchain-openai = "^0.3.18"
langchain-chroma = "^0.1.1"
chromadb = "^0.5.0"
pydantic = "^2.7.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
pyyaml = "^6.0.1"
pydantic-settings = "^2.9.1"
beautifulsoup4 = "^4.12.3"
watchdog = "^6.0.0"
mcp-server-fetch = "^2025.4.7"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
